change the type of vibrance of the vibrance filter to match other filters

+++ framework/Source/GPUImageVibranceFilter.h
@@ -15,6 +15,6 @@
 
 // Modifies the saturation of desaturated colors, leaving saturated colors unmodified.
 // Value -1.2 to 1.2 (-1.2 is minimum vibrance, 0 is no change, and 1.2 is maximum vibrance)
-@property (readwrite, nonatomic) GLfloat vibrance;
+@property (readwrite, nonatomic) CGFloat vibrance;
 
 @end
